refactor, nav, cleaning
