+++ app/src/main/res/drawable/home.xml
@@ -4,14 +4,14 @@
     android:viewportWidth="25"
     android:viewportHeight="25">
     <path
-        android:pathData="M18.867,16.332L18.873,10.539L14.75,6.429C13.506,5.19 11.494,5.19 10.25,6.429L6.136,10.529L6.136,16.329C6.139,18.083 7.564,19.503 9.318,19.5L15.685,19.5C17.438,19.503 18.862,18.085 18.867,16.332Z"
-        android:strokeLineJoin="round"
-        android:strokeWidth="1.5"
         android:fillColor="#00000000"
         android:fillType="evenOdd"
+        android:pathData="M18.867,16.332L18.873,10.539L14.75,6.429C13.506,5.19 11.494,5.19 10.25,6.429L6.136,10.529L6.136,16.329C6.139,18.083 7.564,19.503 9.318,19.5L15.685,19.5C17.438,19.503 18.862,18.085 18.867,16.332Z"
+        android:strokeWidth="1.5"
         android:strokeColor="#000000"
-        android:strokeLineCap="round" />
+        android:strokeLineCap="round"
+        android:strokeLineJoin="round" />
     <path
-        android:pathData="M19.624,6.518C19.624,6.104 19.288,5.768 18.874,5.768C18.46,5.768 18.124,6.104 18.124,6.518L19.624,6.518ZM18.874,10.539L18.124,10.539C18.124,10.738 18.203,10.929 18.344,11.07L18.874,10.539ZM19.971,12.691C20.264,12.984 20.739,12.983 21.031,12.69C21.324,12.396 21.323,11.921 21.029,11.629L19.971,12.691ZM6.666,11.06C6.959,10.768 6.96,10.293 6.667,10C6.375,9.706 5.9,9.705 5.606,9.998L6.666,11.06ZM3.97,11.629C3.677,11.921 3.676,12.396 3.969,12.69C4.261,12.983 4.736,12.984 5.03,12.691L3.97,11.629ZM13.75,19.5C13.75,19.914 14.086,20.25 14.5,20.25C14.914,20.25 15.25,19.914 15.25,19.5L13.75,19.5ZM9.75,19.5C9.75,19.914 10.086,20.25 10.5,20.25C10.914,20.25 11.25,19.914 11.25,19.5L9.75,19.5ZM18.124,6.518L18.124,10.539L19.624,10.539L19.624,6.518L18.124,6.518ZM18.344,11.07L19.971,12.691L21.029,11.629L19.403,10.008L18.344,11.07ZM5.606,9.998L3.97,11.629L5.03,12.691L6.666,11.06L5.606,9.998ZM15.25,19.5L15.25,17.72L13.75,17.72L13.75,19.5L15.25,19.5ZM15.25,17.72C15.25,16.201 14.019,14.97 12.5,14.97L12.5,16.47C13.19,16.47 13.75,17.03 13.75,17.72L15.25,17.72ZM12.5,14.97C10.981,14.97 9.75,16.201 9.75,17.72L11.25,17.72C11.25,17.03 11.81,16.47 12.5,16.47L12.5,14.97ZM9.75,17.72L9.75,19.5L11.25,19.5L11.25,17.72L9.75,17.72Z"
-        android:fillColor="#000000" />
+        android:fillColor="#000000"
+        android:pathData="M19.624,6.518C19.624,6.104 19.288,5.768 18.874,5.768C18.46,5.768 18.124,6.104 18.124,6.518L19.624,6.518ZM18.874,10.539L18.124,10.539C18.124,10.738 18.203,10.929 18.344,11.07L18.874,10.539ZM19.971,12.691C20.264,12.984 20.739,12.983 21.031,12.69C21.324,12.396 21.323,11.921 21.029,11.629L19.971,12.691ZM6.666,11.06C6.959,10.768 6.96,10.293 6.667,10C6.375,9.706 5.9,9.705 5.606,9.998L6.666,11.06ZM3.97,11.629C3.677,11.921 3.676,12.396 3.969,12.69C4.261,12.983 4.736,12.984 5.03,12.691L3.97,11.629ZM13.75,19.5C13.75,19.914 14.086,20.25 14.5,20.25C14.914,20.25 15.25,19.914 15.25,19.5L13.75,19.5ZM9.75,19.5C9.75,19.914 10.086,20.25 10.5,20.25C10.914,20.25 11.25,19.914 11.25,19.5L9.75,19.5ZM18.124,6.518L18.124,10.539L19.624,10.539L19.624,6.518L18.124,6.518ZM18.344,11.07L19.971,12.691L21.029,11.629L19.403,10.008L18.344,11.07ZM5.606,9.998L3.97,11.629L5.03,12.691L6.666,11.06L5.606,9.998ZM15.25,19.5L15.25,17.72L13.75,17.72L13.75,19.5L15.25,19.5ZM15.25,17.72C15.25,16.201 14.019,14.97 12.5,14.97L12.5,16.47C13.19,16.47 13.75,17.03 13.75,17.72L15.25,17.72ZM12.5,14.97C10.981,14.97 9.75,16.201 9.75,17.72L11.25,17.72C11.25,17.03 11.81,16.47 12.5,16.47L12.5,14.97ZM9.75,17.72L9.75,19.5L11.25,19.5L11.25,17.72L9.75,17.72Z" />
 </vector>
