+++ app/src/main/res/layout/activity_main.xml
@@ -16,7 +16,6 @@
         android:orientation="vertical">
 
         <androidx.appcompat.widget.Toolbar
-            xmlns:android="http://schemas.android.com/apk/res/android"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:background="@color/black"
