+++ app/src/main/res/layout/home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FirstFragment">
+    tools:context=".Home">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -75,23 +75,6 @@
                     android:layout_height="wrap_content"
                     android:text="@string/reminder_placeholder" />
             </LinearLayout>
-
-            <!-- development buttons, will remove later -->
-            <Button
-                android:id="@+id/button_first"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/next" />
-            <Button
-                android:id="@+id/button_second"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/next" />
-            <Button
-                android:id="@+id/button_third"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/next" />
         </LinearLayout>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
