+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,9 +7,9 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.medicationapp.FirstFragment"
+        android:name="com.medicationapp.Home"
         android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
+        tools:layout="@layout/home">
